Refactor MVVM
